feat(prisma): add role column to users table
Add a new column named 'role' to the 'users' table in the database. The column is of type integer and has a default value of 1. This change was made to the schema.prisma file and a new migration file was created to add the column to the database. A migration lock file was also added to the project.