refactor(auth.module.ts, user.module.ts): add forwardRef to fix circular dependency issue
The AuthModule and UserModule were dependent on each other, causing a circular dependency issue. To fix this, we added the forwardRef() function to the imports array of both modules. This allows the modules to reference each other without causing a circular dependency issue. Additionally, we added the AuthService to the exports array of the AuthModule to make it available for use in other modules.